Added Checkbox for saving, set icons to static
Icons for controllers are now static untill I figure out how to drag and drop to reorder them.

Checkbox with functionality for saving, now allows users to either save the points of the curve in world position, or from the center of the curve object.

Curves will be created at each selected object, or origo if nothing selected.

Checkbox for adding an offset group now functional.